Pause feature + ship explosion
The ship can now self-destruct with key 'e'.  Later on, we can make asteroids do this.
I am working on adding a pause feature with key 'p'.  We need to make it so the game class updates only specific things in the pause loop at the bottom of the file.